+++ ghc-exactprint.cabal
@@ -31,6 +31,7 @@ library
                      , Language.Haskell.GHC.ExactPrint.Utils
                      , Language.Haskell.GHC.ExactPrint.Common
                      , Language.Haskell.GHC.ExactPrint.Lookup
+                     , Language.Haskell.GHC.ExactPrint.Annotate
   GHC-Options:         -Wall
   -- other-modules:
   -- other-extensions:
@@ -44,6 +45,7 @@ library
                      , mtl
                      , syb
                      , free
+                     , HUnit
   hs-source-dirs:      src
   default-language:    Haskell2010
   -- Note: the following constraint actually requires RC2 or better
