+++ src/Language/Haskell/GHC/ExactPrint/Common.hs
@@ -23,7 +23,7 @@ import Data.List
 
 import Language.Haskell.GHC.ExactPrint.Types
 import Language.Haskell.GHC.ExactPrint.Lookup
-import Language.Haskell.GHC.ExactPrint.Utils (srcSpanStartColumn, srcSpanStartLine, debug, showGhc, isListComp, ss2posEnd, deltaFromSrcSpans, rdrName2String, ss2span, ghcCommentText, isPointSrcSpan, span2ss, ss2deltaP, ghead, undelta)
+import Language.Haskell.GHC.ExactPrint.Utils (srcSpanStartColumn, srcSpanStartLine, debug, showGhc, isListComp, ss2posEnd, deltaFromSrcSpans, rdrName2String, ss2span, ghcCommentText, isPointSrcSpan, span2ss, ss2deltaP, ghead, undelta, isGoodDelta)
 
 import qualified Bag            as GHC
 import qualified BasicTypes     as GHC
@@ -44,9 +44,6 @@ import Control.Monad.Trans.Free
 
 -- ---------------------------------------------------------------------
 
-
-
-data LayoutFlag = LayoutRules | NoLayoutRules deriving (Show, Eq)
 
 data AnnotationF w next where
 --  AddAnnotationWorker :: KeywordId -> GHC.SrcSpan -> next -> AnnotationF w next
@@ -149,106 +146,6 @@ type AnnotateT w m = FreeT (AnnotationF w)  m
 -- ---------------------------------------------------------------------
 
 
-data APState = APState
-             { priorEndPosition :: GHC.SrcSpan
-               -- | Ordered list of comments still to be allocated
-             , apComments :: [Comment]
-               -- | The original GHC API Annotations
-             , apAnns :: GHC.ApiAnns
-             }
-
-data EPState = EPState
-             { epPos       :: Pos -- ^ Current output position
-             , epAnns      :: Anns
-             , epAnnKds    :: [[(KeywordId, DeltaPos)]] -- MP: Could this be moved to the local state with suitable refactoring?
-             }
-
-data StackItem = StackItem
-               { -- | Current `SrcSpan`
-                 curSrcSpan :: GHC.SrcSpan
-                 -- |  `SrcSpan` of the immediately prior scope
-               , prevSrcSpan :: GHC.SrcSpan
-                 -- | The offset required to get from the prior end point to the
-                 -- | The offset required to get from the prior end point to the
-                 -- start of the current SrcSpan. Accessed via `getEntryDP`
-               , offset     :: DeltaPos
-                 -- | Offsets for the elements annotated in this `SrcSpan`
-               -- | Indicates whether the contents of this SrcSpan are
-               -- subject to vertical alignment layout rules
-                 -- | The constructor name of the AST element, to
-                 -- distinguish between nested elements that have the same
-                 -- `SrcSpan` in the AST.
-               , annConName :: AnnConName
-               }
-
-data EPStack = EPStack
-              { epStack     :: (ColOffset,ColDelta) -- ^ stack of offsets that currently apply
-              , epSrcSpan  :: GHC.SrcSpan
-              }
-
-initialEPStack :: GHC.SrcSpan -> EPStack
-initialEPStack ss = EPStack
-    { epStack     = (0,0)
-    , epSrcSpan   = ss
-    }
-
-initialStackItem :: StackItem
-initialStackItem =
-  StackItem
-    { curSrcSpan = GHC.noSrcSpan
-    , prevSrcSpan = GHC.noSrcSpan
-    , offset = DP (0,0)
-    , annConName = annGetConstr ()
-    }
-
-defaultAPState :: GHC.SrcSpan -> GHC.ApiAnns -> APState
-defaultAPState priorEnd ga =
-  let cs = flattenedComments ga in
-    APState
-      { priorEndPosition = priorEnd
-      , apComments = cs
-      , apAnns     = ga    -- $
-      }
-
-defaultEPState :: Anns -> EPState
-defaultEPState as = EPState
-      { epPos    = (1,1)
-      , epAnns   = as
-      , epAnnKds = []
-      }
-
-data APWriter = APWriter
-  { -- Final list of annotations
-    finalAnns :: Endo (Map.Map AnnKey Annotation)
-    -- Used locally to pass Keywords, delta pairs relevant to a specific
-    -- subtree to the parent.
-  , annKds :: [(KeywordId, DeltaPos)]
-    -- Used locally to report a subtrees aderhence to haskell's layout
-    -- rules.
-  , layoutFlag :: LayoutFlag
-  }
-
-
-instance Monoid LayoutFlag where
-  mempty = NoLayoutRules
-  LayoutRules `mappend` _ = LayoutRules
-  _ `mappend` LayoutRules = LayoutRules
-  _ `mappend` _           = NoLayoutRules
-
-tellFinalAnn :: (AnnKey, Annotation) -> AP ()
-tellFinalAnn (k, v) =
-  tell (mempty { finalAnns = Endo (Map.insertWith (<>) k v) })
-
-tellKd :: (KeywordId, DeltaPos) -> AP ()
-tellKd kd = tell (mempty { annKds = [kd] })
-
-setLayoutFlag' :: AP ()
-setLayoutFlag' = tell (mempty {layoutFlag = LayoutRules})
-
-instance Monoid APWriter where
-  mempty = APWriter mempty mempty mempty
-  (APWriter a1 b1 c1) `mappend` (APWriter a2 b2 c2)
-    = APWriter (a1 <> a2) (b1 <> b2) (c1 <> c2)
 
 
 -- | Type used in the AP Monad. The state variables maintain
@@ -259,226 +156,7 @@ instance Monoid APWriter where
 --    - the annotations provided by GHC
 type Wrapped a = AnnotateT (AnnKey, Annotation) Identity a
 
-type AP = RWS StackItem APWriter APState
-
-type EP = RWS EPStack (Endo String) EPState
-
-runAP :: Wrapped () -> GHC.ApiAnns -> GHC.SrcSpan -> Anns
-runAP action ga priorEnd =
-  ($ mempty) . appEndo . finalAnns . snd
-  . (\action -> execRWS action initialStackItem (defaultAPState priorEnd ga))
-  . simpleInterpret $ action
-
-
-runEP :: Wrapped () -> GHC.SrcSpan -> Anns -> String
-runEP action ss ans =
-  flip appEndo "" . snd
-  . (\action -> execRWS action (initialEPStack ss) (defaultEPState ans))
-  . printInterpret $ action
-
-simpleInterpret :: AnnotateT (AnnKey, Annotation) Identity a -> AP a
-simpleInterpret = iterTM go
-  where
-    go :: AnnotationF (AnnKey, Annotation) (AP a) -> AP a
-    go (Output w next) = next
-    go (AddEofAnnotation next) = addEofAnnotation' >> next
-    go (AddDeltaAnnotation kwid next) =
-      addDeltaAnnotation' kwid >> next
-    go (AddDeltaAnnotationsOutside akwid kwid next) = addDeltaAnnotationsOutside' akwid kwid >> next
-    go (AddDeltaAnnotationsInside akwid next) = addDeltaAnnotationsInside' akwid >> next
-    go (AddDeltaAnnotations akwid next) = addDeltaAnnotations' akwid >> next
-    go (AddDeltaAnnotationLs akwid n next) = addDeltaAnnotationLs' akwid n >> next
-    go (AddDeltaAnnotationAfter akwid next) = addDeltaAnnotationAfter' akwid >> next
-    go (AddDeltaAnnotationExt ss akwid next) = addDeltaAnnotationExt' ss akwid >> next
-    go (WithAST lss layoutflag prog next) =
-      withAST' lss layoutflag (simpleInterpret prog) >>= next
-    go (OutputKD (kwid, (_, dp)) next) = tellKd (dp, kwid) >> next
-    go (CountAnnsAP kwid next) = countAnnsAP' kwid >>= next
-    go (SetLayoutFlag next) = setLayoutFlag' >> next
-    go (PrintAnnString akwid s next) = addDeltaAnnotation' akwid >> next
-    go (PrintAnnStringExt ss akwid _ next) = addDeltaAnnotationExt' ss akwid >> next
-    go (PrintAnnStringLs akwid _ n next) = addDeltaAnnotationLs' akwid n >> next
-
-beforeAction :: GHC.GenLocated GHC.SrcSpan e
-                 -> (GHC.SrcSpan
-                     -> DeltaPos
-                     -> GHC.SrcSpan
-                     -> DeltaPos
-                     -> AP b)
-                 -> AP b
-beforeAction lss act = do
-  pe <- getPriorEnd
-  let ss = (GHC.getLoc lss)
-  let edp = deltaFromSrcSpans pe ss
-  edp' <- adjustDeltaForOffsetM edp
-  act pe edp' ss edp
-
---middleAction :: GHC.Located a -> DeltaPos -> AP b -> _ -> AP c
-middleAction :: Data a => GHC.Located a -> DeltaPos -> AP b -> ((b, APWriter) -> AP c) -> AP c
-middleAction lss edp wrapped k = censor maskWriter $ do
-  withSrcSpanAP lss edp (listen wrapped) >>= k
-  where
-    maskWriter s = s { annKds = []
-                     , layoutFlag = NoLayoutRules }
-
-afterAction :: LayoutFlag -> GHC.SrcSpan -> [(KeywordId, DeltaPos)] -> AP ()
-afterAction lf ss annkds = do
-  (dp,nl)  <- getCurrentDP lf
-  finaledp <- getEntryDP
-  addAnnotationsAP (Ann finaledp nl (srcSpanStartColumn ss) dp annkds)
-
-printInterpret :: AnnotateT (AnnKey, Annotation) Identity a -> EP a
-printInterpret = iterTM go
-  where
-    go :: AnnotationF (AnnKey, Annotation) (EP a) -> EP a
-    go (Output w next) = next
-    go (AddEofAnnotation next) = printStringAtMaybeAnn (G GHC.AnnEofPos) "" >> next
-    go (AddDeltaAnnotation kwid next) =
-      justOne kwid  >> next
-    go (AddDeltaAnnotationsOutside akwid kwid next) =
-      printStringAtMaybeAnnAll kwid ";"  >> next
-    go (AddDeltaAnnotationsInside akwid next) =
-      allAnns akwid >> next
-    go (AddDeltaAnnotations akwid next) =
-      allAnns akwid >> next
-    go (AddDeltaAnnotationLs akwid _ next) =
-      justOne akwid >> next
-    go (AddDeltaAnnotationAfter akwid next) =
-      justOne akwid >> next
-    go (AddDeltaAnnotationExt ss akwid next) =
-      justOne akwid >> next
-    go (WithAST lss layoutflag action next) =
-      exactPC lss (printInterpret action) >>= next
-    go (OutputKD _ next) =
-      next
-    go (CountAnnsAP kwid next) =
-      countAnnsEP (G kwid) >>= next
-    go (SetLayoutFlag next) =
-      next
-    go (PrintAnnString akwid s next) = printStringAtMaybeAnn (G akwid) s >> next
-    go (PrintAnnStringExt _ akwid s next) = printStringAtMaybeAnn (G akwid) s >> next
-    go (PrintAnnStringLs akwid s _ next) = printStringAtMaybeAnn (G akwid) s >> next
-
-justOne kwid = printStringAtMaybeAnn (G kwid) (keywordToString kwid)
-allAnns kwid = printStringAtMaybeAnnAll (G kwid) (keywordToString kwid)
-
-
 -- ---------------------------------------------------------------------
-
-flattenedComments :: GHC.ApiAnns -> [Comment]
-flattenedComments (_,cm) = map tokComment . GHC.sortLocated . concat $ Map.elems cm
-
--- -------------------------------------
-
-getSrcSpanAP :: AP GHC.SrcSpan
-getSrcSpanAP = asks curSrcSpan
-
-getPriorSrcSpanAP :: AP GHC.SrcSpan
-getPriorSrcSpanAP = asks prevSrcSpan
-
-withSrcSpanAP :: Data a => (GHC.Located a) -> DeltaPos -> AP b -> AP b
-withSrcSpanAP (GHC.L l a) edp =
-  local (\s -> let previousSrcSpan = curSrcSpan s in
-               s { curSrcSpan = l
-                 , prevSrcSpan = previousSrcSpan
-                 , offset = edp
-                 , annConName = annGetConstr a
-                 })
-
-
-getEntryDP :: AP DeltaPos
-getEntryDP = asks offset
-
-getUnallocatedComments :: AP [Comment]
-getUnallocatedComments = gets apComments
-
-putUnallocatedComments :: [Comment] -> AP ()
-putUnallocatedComments cs = modify (\s -> s { apComments = cs } )
-
--- ---------------------------------------------------------------------
-
-adjustDeltaForOffsetM :: DeltaPos -> AP DeltaPos
-adjustDeltaForOffsetM dp = do
-  colOffset <- getCurrentColOffset
-  return (adjustDeltaForOffset colOffset dp)
-
-adjustDeltaForOffset :: Int -> DeltaPos -> DeltaPos
-adjustDeltaForOffset _colOffset dp@(DP (0,_)) = dp -- same line
-adjustDeltaForOffset  colOffset    (DP (l,c)) = DP (l,c - colOffset)
-
--- ---------------------------------------------------------------------
-
--- | Get the current column offset
-getCurrentColOffset :: AP ColOffset
-getCurrentColOffset = srcSpanStartColumn <$> getSrcSpanAP
-
--- |Get the difference between the current and the previous
--- colOffsets, if they are on the same line
-getCurrentDP :: LayoutFlag -> AP (ColOffset,LineChanged)
-getCurrentDP layoutOn = do
-  -- Note: the current col offsets are not needed here, any
-  -- indentation should be fully nested in an AST element
-  ss <- getSrcSpanAP
-  ps <- getPriorSrcSpanAP
-  let boolLayoutFlag = case layoutOn of { LayoutRules -> True; NoLayoutRules -> False}
-      colOffset = if srcSpanStartLine ss == srcSpanStartLine ps
-                    then srcSpanStartColumn ss - srcSpanStartColumn ps
-                    else srcSpanStartColumn ss
-      r = case ( boolLayoutFlag , srcSpanStartLine ss == srcSpanStartLine ps) of
-             (True,  True) -> (colOffset, LayoutLineSame)
-             (True, False) -> (colOffset, LayoutLineChanged)
-             (False, True) -> (colOffset, LineSame)
-             (False,False) -> (colOffset, LineChanged)
-  return r
-    `debug` ("getCurrentDP:layoutOn=" ++ show layoutOn)
-
--- ---------------------------------------------------------------------
-
--- |Note: assumes the prior end SrcSpan stack is nonempty
-getPriorEnd :: AP GHC.SrcSpan
-getPriorEnd = gets priorEndPosition
-
-setPriorEnd :: GHC.SrcSpan -> AP ()
-setPriorEnd pe = modify (\s -> s { priorEndPosition = pe })
-
--- -------------------------------------
-
-getAnnotationAP :: GHC.AnnKeywordId -> AP [GHC.SrcSpan]
-getAnnotationAP an = do
-    ga <- gets apAnns
-    ss <- getSrcSpanAP
-    return $ GHC.getAnnotation ga ss an
-
-getAndRemoveAnnotationAP :: GHC.SrcSpan -> GHC.AnnKeywordId -> AP [GHC.SrcSpan]
-getAndRemoveAnnotationAP sp an = do
-    ga <- gets apAnns
-    let (r,ga') = GHC.getAndRemoveAnnotation ga sp an
-    r <$ modify (\s -> s { apAnns = ga' })
-
-
-tokComment :: GHC.Located GHC.AnnotationComment -> Comment
-tokComment t@(GHC.L lt _) = Comment (ss2span lt) (ghcCommentText t)
-
--- -------------------------------------
-
--- |Add some annotation to the currently active SrcSpan
-addAnnotationsAP :: Annotation -> AP ()
-addAnnotationsAP ann = do
-    l <- ask
-    tellFinalAnn ((getAnnKey l),ann)
-
-getAnnKey :: StackItem -> AnnKey
-getAnnKey StackItem {curSrcSpan, annConName} = AnnKey curSrcSpan annConName
-
--- -------------------------------------
-
-addAnnDeltaPos :: (GHC.SrcSpan,KeywordId) -> DeltaPos -> AP ()
-addAnnDeltaPos (_s,kw) dp = tellKd (kw, dp)
-
--- -------------------------------------
-
---withAST' :: Data a => GHC.Located a -> LayoutFlag -> Wrapped b -> Wrapped b
---withAST' l lf action  = join . liftF $ Recurse l lf action (return ())
 
 withAST :: Data a => GHC.Located a -> LayoutFlag -> Wrapped b -> Wrapped b
 withAST lss layout action = liftF (WithAST lss layout prog (\b -> b))
@@ -489,316 +167,6 @@ withAST lss layout action = liftF (WithAST lss layout prog (\b -> b))
       addDeltaAnnotationAfter GHC.AnnComma
       addDeltaAnnotationsOutside GHC.AnnSemi AnnSemiSep
       return r
-
--- | Enter a new AST element. Maintain SrcSpan stack
-withAST' :: Data a => GHC.Located a -> LayoutFlag -> AP b -> AP b
-withAST' lss layout action = do
-  -- Calculate offset required to get to the start of the SrcSPan
-  pe <- getPriorEnd
-  let ss = (GHC.getLoc lss)
-  let edp = deltaFromSrcSpans pe ss
-  edp' <- adjustDeltaForOffsetM edp
-  -- need to save edp', and put it in Annotation
-
-  withSrcSpanAP lss edp' (do
-
-    let maskWriter s = s { annKds = []
-                         , layoutFlag = NoLayoutRules }
-
-    (res, w) <-
-      (censor maskWriter (listen action))
-
-    (dp,nl)  <- getCurrentDP (layout <> layoutFlag w)
-    finaledp <- getEntryDP
-    let kds = annKds w
-    addAnnotationsAP (Ann finaledp nl (srcSpanStartColumn ss) dp kds)
-      `debug` ("leaveAST:(ss,edp,dp,kds)=" ++ show (showGhc ss,edp,dp,kds,dp))
-    return res)
-
-{-
-withAST' :: Data a => GHC.Located a -> LayoutFlag -> Wrapped b -> Wrapped b
-withAST' lss@(GHC.L l ast) layout action = do
-
-  -- Calculate offset required to get to the start of the SrcSPan
-
-  pe <- getPriorEnd
-  let ss = (GHC.getLoc lss)
-  let edp = deltaFromSrcSpans pe ss
-  edp' <- adjustDeltaForOffsetM edp
-  -- need to save edp', and put it in Annotation
-  --(withContext kds l an . withSrcSpanAP lss edp') (do
-
-
-  (res, w) <-
-   middle lss edp'
-     (do
-        r <- action
-        -- Automatically add any trailing comma or semi
-        addDeltaAnnotationAfter GHC.AnnComma
-        addDeltaAnnotationsOutside GHC.AnnSemi AnnSemiSep
-        return r)
-  after res layout (layoutFlag w) ss (annKds w)
-    {-
-    return res)
-    -}
--}
--- |First move to the given location, then call exactP
-exactPC :: Data ast => GHC.Located ast -> EP a -> EP a
-exactPC a@(GHC.L l ast) action =
-    do return () `debug` ("exactPC entered for:" ++ showGhc l)
-       ma <- getAndRemoveAnnotation a
-       let an@(Ann _edp _nl _sc _dc kds) = fromMaybe annNone ma
-       withContext kds l an action
-
-withContext :: [(KeywordId, DeltaPos)]
-            -> GHC.SrcSpan
-            -> Annotation
-            -> EP a -> EP a
-withContext kds l an = withKds kds . withSrcSpan l . withOffset an
-
-
-getPos :: EP Pos
-getPos = gets epPos
-
-setPos :: Pos -> EP ()
-setPos l = modify (\s -> s {epPos = l})
-
--- ---------------------------------------------------------------------
-
--- | Given an annotation associated with a specific SrcSpan, determines a new offset relative to the previous
--- offset
---
-withOffset :: Annotation -> EP a -> EP a
-withOffset a@(Ann (DP (edLine, edColumn)) newline originalStartCol annDelta _) k = do
-  -- The colOffset is the offset to be used when going to the next line.
-  -- The colIndent is how far the current code block has been indented.
-  (colOffset, colIndent) <- asks epStack
-  (_l, currentColumn) <- getPos
-  let
-      -- Add extra indentation for this SrcSpan
-     colOffset' = annDelta + colOffset
-     -- Work out where new column starts, based on the entry delta
-     newStartColumn = if edLine == 0
-                          then edColumn + currentColumn -- same line, use entry delta
-                          else colOffset -- different line, use current offset
-     newColIndent        = newStartColumn - originalStartCol
-
-     offsetValNewIndent  = (colOffset' + (newColIndent - colIndent), newColIndent)
-     offsetValSameIndent = (colOffset'            ,                  colIndent)
-     offsetValNewline    = (annDelta   + colIndent,                  colIndent)
-
-     newOffset =
-        case newline of
-          -- For use by AST modifiers, to preserve the indentation
-          -- level for the next line after an AST modification
-          KeepOffset  -> offsetValNewline
-
-          -- Generated during the annotation phase
-          LineChanged       -> offsetValNewline
-          LayoutLineChanged -> offsetValNewline
-
-          LineSame          -> offsetValSameIndent
-          LayoutLineSame    -> offsetValNewIndent
-  local (\s -> s {epStack = newOffset }) k
-    `debug` ("pushOffset:(a, colOffset, colIndent, currentColumn, newOffset)="
-                 ++ show (a, colOffset, colIndent, currentColumn, newOffset))
-
--- |Get the current column offset
-getOffset :: EP ColOffset
-getOffset = asks (fst . epStack)
-
--- ---------------------------------------------------------------------
-
-withSrcSpan :: GHC.SrcSpan -> EP a -> EP a
-withSrcSpan ss = local (\s -> s {epSrcSpan = ss})
-
-getAndRemoveAnnotation :: (Data a) => GHC.Located a -> EP (Maybe Annotation)
-getAndRemoveAnnotation a = do
-  (r, an') <- gets (getAndRemoveAnnotationEP a . epAnns)
-  modify (\s -> s { epAnns = an' })
-  return r
-
-withKds :: [(KeywordId, DeltaPos)] -> EP a -> EP a
-withKds kd action = do
-  modify (\s -> s { epAnnKds = kd : (epAnnKds s) })
-  r <- action
-  modify (\s -> s { epAnnKds = tail (epAnnKds s) })
-  return r
-
--- | Get and remove the first item in the (k,v) list for which the k matches.
--- Return the value, together with any comments skipped over to get there.
-destructiveGetFirst :: KeywordId -> ([(KeywordId,v)],[(KeywordId,v)])
-                    -> ([(KeywordId,v)], Maybe v,[(KeywordId,v)])
-destructiveGetFirst _key (acc,[]) = ([], Nothing ,acc)
-destructiveGetFirst  key (acc,((k,v):kvs))
-  | k == key = let (cs,others) = commentsAndOthers acc in (cs, Just v ,others++kvs)
-  | otherwise = destructiveGetFirst key (acc++[(k,v)],kvs)
-  where
-    commentsAndOthers kvs' = partition isComment kvs'
-    isComment ((AnnComment _),_) = True
-    isComment _              = False
-
--- |destructive get, hence use an annotation once only
-getAnnFinal :: KeywordId -> EP ([DComment], Maybe DeltaPos)
-getAnnFinal kw = do
-  kd <- gets epAnnKds
-  let (r, kd', dcs) = case kd of
-                  []    -> (Nothing ,[], [])
-                  (k:kds) -> (r',kk:kds, dcs')
-                    where (cs', r',kk) = destructiveGetFirst kw ([],k)
-                          dcs' = concatMap keywordIdToDComment cs'
-  modify (\s -> s { epAnnKds = kd' })
-  return (dcs, r)
-
--- ---------------------------------------------------------------------
-
-getStoredListSrcSpan :: EP GHC.SrcSpan
-getStoredListSrcSpan = do
-  kd <- gets epAnnKds
-  let
-    isAnnList ((AnnList _),_) = True
-    isAnnList _               = False
-
-    kdf = ghead "getStoredListSrcSpan.1" kd
-    (AnnList ss,_) = ghead "getStoredListSrcSpan.2" $ filter isAnnList kdf
-  return ss
-
--- ---------------------------------------------------------------------
-
-keywordIdToDComment :: (KeywordId, DeltaPos) -> [DComment]
-keywordIdToDComment (AnnComment comment,_dp) = [comment]
-keywordIdToDComment _                   = []
-
--- |non-destructive get
-peekAnnFinal :: KeywordId -> EP (Maybe DeltaPos)
-peekAnnFinal kw = do
-  (_, r, _) <- (\kd -> destructiveGetFirst kw ([], kd)) <$> gets (head . epAnnKds)
-  return r
--- ---------------------------------------------------------------------
-printString :: String -> EP ()
-printString str = do
-  (l,c) <- gets epPos
-  setPos (l, c + length str)
-  tell (Endo $ showString str )
-
-newLine :: EP ()
-newLine = do
-    (l,_) <- getPos
-    printString "\n"
-    setPos (l+1,1)
-
-padUntil :: Pos -> EP ()
-padUntil (l,c) = do
-    (l1,c1) <- getPos
-    case  {- trace (show ((l,c), (l1,c1))) -} () of
-     _ {-()-} | l1 >= l && c1 <= c -> printString $ replicate (c - c1) ' '
-              | l1 < l             -> newLine >> padUntil (l,c)
-              | otherwise          -> return ()
-
-printWhitespace :: Pos -> EP ()
-printWhitespace p = do
-  -- mPrintComments p >> padUntil p
-  padUntil p
-
-printStringAt :: Pos -> String -> EP ()
-printStringAt p str = printWhitespace p >> printString str
-
--- ---------------------------------------------------------------------
-
--- |This should be the final point where things are mode concrete,
--- before output. Hence the point where comments can be inserted
-printStringAtLsDelta :: [DComment] -> [DeltaPos] -> String -> EP ()
-printStringAtLsDelta cs mc s =
-  case reverse mc of
-    (cl:_) -> do
-      p <- getPos
-      colOffset <- getOffset
-      if isGoodDeltaWithOffset cl colOffset
-        then do
-          mapM_ printQueuedComment cs
-          printStringAt (undelta p cl colOffset) s
-            `debug` ("printStringAtLsDelta:(pos,s):" ++ show (undelta p cl colOffset,s))
-        else return () `debug` ("printStringAtLsDelta:bad delta for (mc,s):" ++ show (mc,s))
-    _ -> return ()
-
-
-isGoodDeltaWithOffset :: DeltaPos -> Int -> Bool
-isGoodDeltaWithOffset dp colOffset = isGoodDelta (DP (undelta (0,0) dp colOffset))
-
--- AZ:TODO: harvest the commonality between this and printStringAtLsDelta
-printQueuedComment :: DComment -> EP ()
-printQueuedComment (DComment (dp,de) s) = do
-  p <- getPos
-  colOffset <- getOffset
-  let (dr,dc) = undelta (0,0) dp colOffset
-  if isGoodDelta (DP (dr,max 0 dc)) -- do not lose comments against the left margin
-    then do
-      printStringAt (undelta p dp colOffset) s
-         `debug` ("printQueuedComment:(pos,s):" ++ show (undelta p dp colOffset,s))
-      setPos (undelta p de colOffset)
-    else return () `debug` ("printQueuedComment::bad delta for (dp,s):" ++ show (dp,s))
--- ---------------------------------------------------------------------
-
-printStringAtMaybeAnn :: KeywordId -> String -> EP ()
-printStringAtMaybeAnn an str = do
-  (comments, ma) <- getAnnFinal an
-  printStringAtLsDelta comments (maybeToList ma) str
-    `debug` ("printStringAtMaybeAnn:(an,ma,str)=" ++ show (an,ma,str))
-
-printStringAtMaybeAnnAll :: KeywordId -> String -> EP ()
-printStringAtMaybeAnnAll an str = go
-  where
-    go = do
-      (comments, ma) <- getAnnFinal an
-      case ma of
-        Nothing -> return ()
-        Just d  -> printStringAtLsDelta comments [d] str >> go
--- ---------------------------------------------------------------------
-
-countAnnsEP :: KeywordId -> EP Int
-countAnnsEP an = do
-  ma <- peekAnnFinal an
-  return (length ma)
-------------------------------------------------------------------------------
--- Printing of source elements
--- | Print an AST exactly as specified by the annotations on the nodes in the tree.
--- exactPrint :: (ExactP ast) => ast -> [Comment] -> String
-exactPrint :: (AnnotateGen ast) => GHC.Located ast -> String
-exactPrint ast@(GHC.L l _) = runEP (annotatePC ast) l Map.empty
-
-
-exactPrintAnnotated ::
-     GHC.Located (GHC.HsModule GHC.RdrName) -> GHC.ApiAnns -> String
-exactPrintAnnotated ast@(GHC.L l _) ghcAnns = runEP (annotatePC ast) l an
-  where
-    an = annotateLHsModule ast ghcAnns
-
-exactPrintAnnotation :: AnnotateGen ast =>
-  GHC.Located ast -> Anns -> String
-exactPrintAnnotation ast@(GHC.L l _) an = runEP (annotatePC ast) l an
-  -- `debug` ("exactPrintAnnotation:an=" ++ (concatMap (\(l,a) -> show (ss2span l,a)) $ Map.toList an ))
-  --
-
-annotateAST :: GHC.Located (GHC.HsModule GHC.RdrName) -> GHC.ApiAnns -> Anns
-annotateAST ast ghcAnns = annotateLHsModule ast ghcAnns
-
-{-
--- ---------------------------------------------------------------------
-
--- ---------------------------------------------------------------------
-
-printMerged :: (AnnotateGen a, AnnotateGen b) => [GHC.Located a] -> [GHC.Located b] -> Wrapped  ()
-printMerged [] [] = return ()
-printMerged [] bs = mapM_ annotatePC bs
-printMerged as [] = mapM_ annotatePC as
-printMerged (a@(GHC.L l1 _):as) (b@(GHC.L l2 _):bs) =
-  if l1 < l2
-    then annotatePC a >> printMerged    as (b:bs)
-    else annotatePC b >> printMerged (a:as)   bs
-
--}
--- ---------------------------------------------------------------------
-
 
 -- ---------------------------------------------------------------------
 
@@ -831,159 +199,10 @@ annotateListWithLayout l ls = do
 
 -- ---------------------------------------------------------------------
 
-isGoodDelta :: DeltaPos -> Bool
-isGoodDelta (DP (ro,co)) = ro >= 0 && co >= 0
+-- ---------------------------------------------------------------------
 
 -- ---------------------------------------------------------------------
 
--- |Split the ordered list of comments into ones that occur prior to
--- the given SrcSpan and the rest
-allocatePriorComments :: [Comment] -> GHC.SrcSpan -> ([Comment],[Comment])
-allocatePriorComments cs ss = partition isPrior cs
-  where
-    (start,_) = ss2span ss
-    isPrior (Comment s _)  = (fst s) < start
-      `debug` ("allocatePriorComments:(s,ss,cond)=" ++ showGhc (s,ss,(fst s) < start))
-
--- ---------------------------------------------------------------------
-
-addAnnotationWorker :: KeywordId -> GHC.SrcSpan -> AP ()
-addAnnotationWorker ann pa = do
-  if not (isPointSrcSpan pa)
-    then do
-      pe <- getPriorEnd
-      ss <- getSrcSpanAP
-      let p = deltaFromSrcSpans pe pa
-      case (ann,isGoodDelta p) of
-        (G GHC.AnnComma,False) -> return ()
-        (G GHC.AnnSemi, False) -> return ()
-        (G GHC.AnnOpen, False) -> return ()
-        (G GHC.AnnClose,False) -> return ()
-        _ -> do
-          cs <- getUnallocatedComments
-          let (allocated,cs') = allocatePriorComments cs pa
-          putUnallocatedComments cs'
-          return () `debug`("addAnnotationWorker:(ss,pa,allocated,cs)=" ++ showGhc (ss,pa,allocated,cs))
-          mapM_ addDeltaComment allocated
-          p' <- adjustDeltaForOffsetM p
-          addAnnDeltaPos (ss,ann) p'
-          setPriorEnd pa
-              -- `debug` ("addDeltaAnnotationWorker:(ss,pe,pa,p,ann)=" ++ show (ss2span ss,ss2span pe,ss2span pa,p,ann))
-    else do
-      return ()
-          -- `debug` ("addDeltaAnnotationWorker::point span:(ss,ma,ann)=" ++ show (ss2span pa,ann))
-
--- ---------------------------------------------------------------------
-
-addDeltaComment :: Comment -> AP ()
-addDeltaComment (Comment paspan str) = do
-  let pa = span2ss paspan
-  pe <- getPriorEnd
-  ss <- getSrcSpanAP
-  let p = deltaFromSrcSpans pe pa
-  p' <- adjustDeltaForOffsetM p
-  setPriorEnd pa
-  let e = ss2deltaP (ss2posEnd pe) (snd paspan)
-  e' <- adjustDeltaForOffsetM e
-  addAnnDeltaPos (ss,AnnComment (DComment (p',e') str)) p'
-
--- ---------------------------------------------------------------------
-
--- | Look up and add a Delta annotation at the current position, and
--- advance the position to the end of the annotation
-addDeltaAnnotation' :: GHC.AnnKeywordId -> AP ()
-addDeltaAnnotation' ann = do
-  ss <- getSrcSpanAP
-  ma <- getAnnotationAP ann
-  case nub ma of -- ++AZ++ TODO: get rid of duplicates earlier
-    [] -> return () `debug` ("addDeltaAnnotation empty ma for:" ++ show ann)
-    [pa] -> addAnnotationWorker (G ann) pa
-    _ -> error $ "addDeltaAnnotation:(ss,ann,ma)=" ++ showGhc (ss,ann,ma)
-
--- | Look up and add a Delta annotation appearing beyond the current
--- SrcSpan at the current position, and advance the position to the
--- end of the annotation
-addDeltaAnnotationAfter' :: GHC.AnnKeywordId -> AP ()
-addDeltaAnnotationAfter' ann = do
-  ss <- getSrcSpanAP
-  ma <- getAnnotationAP ann
-  let ma' = filter (\s -> not (GHC.isSubspanOf s ss)) ma
-  case ma' of
-    [] -> return () `debug` ("addDeltaAnnotation empty ma")
-    [pa] -> addAnnotationWorker (G ann) pa
-    _ -> error $ "addDeltaAnnotation:(ss,ann,ma)=" ++ showGhc (ss,ann,ma)
-
--- | Look up and add a Delta annotation at the current position, and
--- advance the position to the end of the annotation
-addDeltaAnnotationLs' :: GHC.AnnKeywordId -> Int -> AP ()
-addDeltaAnnotationLs' ann off = do
-  ma <- getAnnotationAP ann
-  case (drop off ma) of
-    [] -> return ()
-        -- `debug` ("addDeltaAnnotationLs:missed:(off,pe,ann,ma)=" ++ show (off,ss2span pe,ann,fmap ss2span ma))
-    (pa:_) -> addAnnotationWorker (G ann) pa
-
--- | Look up and add possibly multiple Delta annotation at the current
--- position, and advance the position to the end of the annotations
-addDeltaAnnotations' :: GHC.AnnKeywordId -> AP ()
-addDeltaAnnotations' ann = do
-  ma <- getAnnotationAP ann
-  let do_one ap' = addAnnotationWorker (G ann) ap'
-                    -- `debug` ("addDeltaAnnotations:do_one:(ap',ann)=" ++ showGhc (ap',ann))
-  mapM_ do_one (sort ma)
-
--- | Look up and add possibly multiple Delta annotations enclosed by
--- the current SrcSpan at the current position, and advance the
--- position to the end of the annotations
-addDeltaAnnotationsInside' :: GHC.AnnKeywordId -> AP ()
-addDeltaAnnotationsInside' ann = do
-  ss <- getSrcSpanAP
-  ma <- getAnnotationAP ann
-  let do_one ap' = addAnnotationWorker (G ann) ap'
-                    -- `debug` ("addDeltaAnnotations:do_one:(ap',ann)=" ++ showGhc (ap',ann))
-  mapM_ do_one (sort $ filter (\s -> GHC.isSubspanOf s ss) ma)
-
--- | Look up and add possibly multiple Delta annotations not enclosed by
--- the current SrcSpan at the current position, and advance the
--- position to the end of the annotations
-addDeltaAnnotationsOutside' :: GHC.AnnKeywordId -> KeywordId -> AP ()
-addDeltaAnnotationsOutside' gann ann = do
-  ss <- getSrcSpanAP
-  if ss2span ss == ((1,1),(1,1))
-    then return ()
-    else do
-      -- ma <- getAnnotationAP ss gann
-      ma <- getAndRemoveAnnotationAP ss gann
-      let do_one ap' = addAnnotationWorker ann ap'
-                        -- `debug` ("addDeltaAnnotations:do_one:(ap',ann)=" ++ showGhc (ap',ann))
-      mapM_ do_one (sort $ filter (\s -> not (GHC.isSubspanOf s ss)) ma)
-
--- | Add a Delta annotation at the current position, and advance the
--- position to the end of the annotation
-addDeltaAnnotationExt' :: GHC.SrcSpan -> GHC.AnnKeywordId -> AP ()
-addDeltaAnnotationExt' s ann = addAnnotationWorker (G ann) s
-
-
-
-addEofAnnotation' :: AP ()
-addEofAnnotation' = do
-  pe <- getPriorEnd
-  ss <- getSrcSpanAP
-  ma <- withSrcSpanAP (GHC.noLoc ()) (DP (0,0)) (getAnnotationAP GHC.AnnEofPos)
-  case ma of
-    [] -> return ()
-    (pa:pss) -> do
-      cs <- getUnallocatedComments
-      mapM_ addDeltaComment cs
-      let DP (r,c) = deltaFromSrcSpans pe pa
-      addAnnDeltaPos (ss,G GHC.AnnEofPos) (DP (r, c - 1))
-      setPriorEnd pa
-
-
-countAnnsAP' :: GHC.AnnKeywordId -> AP Int
-countAnnsAP' ann = do
-  ma <- getAnnotationAP ann
-  return (length ma)
 
 -- ---------------------------------------------------------------------
 -- Managing lists which have been separated, e.g. Sigs and Binds
@@ -999,11 +218,6 @@ applyListAnnotations ls
 -- Start of application specific part
 
 -- ---------------------------------------------------------------------
-
-annotateLHsModule :: GHC.Located (GHC.HsModule GHC.RdrName) -> GHC.ApiAnns
-                  -> Anns
-annotateLHsModule modu@(GHC.L ss _) ghcAnns
-   = runAP (annotatePC modu) ghcAnns ss
 
 -- ---------------------------------------------------------------------
